pyproject.toml
[project]
name = "open-macromolecular-genome-polyagent"
version = "0.1.0"
description = "generates polymer smiles for polyagent"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    
    "mcp[cli]>=1.21.0",
    "torch>=1.3.0",
    "numpy>=1.17.2,<2.0",
    "tqdm>=4.36.1",
    "absl-py>=1.4.0",
    "arrow>=1.2.3",
    "rdkit-pypi>=2022.9.5",
    "pydantic>=2.12.4",
    "pandas>=2.3.3",
    "scikit-learn>=1.7.2",
    "matplotlib>=3.10.7",
    "optuna>=4.6.0",
    "seaborn>=0.13.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=5.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
[tool.setuptools.packages.find]
where = ["."]
include = ["OpenMacromolecularGenome*"]
